Split beams_input into several smaller files
