+++ src/restruct_poc/CMakeLists.txt
@@ -1,4 +1,3 @@
 target_sources(${oturb_lib_name}
     PRIVATE
-    beams_input.cpp
 )
