+++ tests/unit_tests/restruct_poc/test_cantilever_beam.cpp
@@ -5,11 +5,18 @@
 #include <gtest/gtest.h>
 
 #include "src/restruct_poc/beams.hpp"
-#include "src/restruct_poc/beams_input.hpp"
+#include "src/restruct_poc/BeamSection.hpp"
+#include "src/restruct_poc/BeamNode.hpp"
+#include "src/restruct_poc/BeamElement.hpp"
+#include "src/restruct_poc/BeamsInput.hpp"
+#include "src/restruct_poc/CreateBeams.hpp"
 #include "src/restruct_poc/Solver.hpp"
 #include "src/restruct_poc/InitializeConstraints.hpp"
 #include "src/restruct_poc/Step.hpp"
 #include "tests/unit_tests/gen_alpha_poc/test_utilities.h"
+
+using BeamQuadrature = std::vector<std::array<double, 2>>;
+using Array_6x6 = std::array<std::array<double, 6>, 6>;
 
 namespace openturbine::restruct_poc::tests {
